fix failing workspace quality checks
