+++ packages/server/src/api/workspaces.js
@@ -132,6 +132,60 @@ function handleCreateError(res, session, error, label) {
   return res.status(500).json({ error: error.message || 'Internal server error' });
 }
 
+function parseWorkspaceCardOptions({ archived, starred, limit, offset, status, scheduled }) {
+  const parsedLimit = Number.parseInt(limit, 10);
+  const parsedOffset = offset === undefined ? 0 : Number.parseInt(offset, 10);
+  const valid = Number.isInteger(parsedLimit) && parsedLimit >= 1 && parsedLimit <= 50
+    && Number.isInteger(parsedOffset) && parsedOffset >= 0
+    && ['running', 'idle', undefined].includes(status)
+    && ['true', 'false', undefined].includes(scheduled);
+  if (!valid) return null;
+  return {
+    archived: archived === 'true',
+    starred: starred === 'true' ? true : starred === 'false' ? false : null,
+    status: status || null,
+    scheduled: scheduled === 'true' ? true : scheduled === 'false' ? false : null,
+    limit: parsedLimit,
+    offset: parsedOffset,
+  };
+}
+
+function sendWorkspaceCards(res, projectId, query, startedAt) {
+  const options = parseWorkspaceCardOptions(query);
+  if (!options) return res.status(400).json({ error: 'Invalid workspace card pagination or filters' });
+  const pagePlusOne = sessions.getWorkspaceCards(projectId, { ...options, limit: options.limit + 1 });
+  return sendWorkspaceJson(res, {
+    workspaces: pagePlusOne.slice(0, options.limit),
+    pagination: { limit: options.limit, offset: options.offset, hasMore: pagePlusOne.length > options.limit },
+  }, startedAt);
+}
+
+function listProjectWorkspaces(req, res) {
+  const startedAt = performance.now();
+  const project = projects.getById(req.params.projectId);
+  if (!project) return res.status(404).json({ error: ERR_PROJECT_NOT_FOUND });
+
+  const { archived, starred, limit, offset, view } = req.query;
+  if (view === 'cards') return sendWorkspaceCards(res, req.params.projectId, req.query, startedAt);
+
+  const archivedFilter = archived === 'true' ? true : archived === 'false' ? false : null;
+  const starredFilter = starred === 'true' ? true : starred === 'false' ? false : null;
+  const parsedLimit = limit ? parseInt(limit, 10) : null;
+  const parsedOffset = offset ? parseInt(offset, 10) : 0;
+  const workspaces = sessions.getRootsByProjectId(req.params.projectId, {
+    archived: archivedFilter, starred: starredFilter, limit: parsedLimit, offset: parsedOffset,
+  });
+  if (parsedLimit === null) return res.json(workspaces);
+
+  const total = sessions.getRootsCountByProjectId(req.params.projectId, {
+    archived: archivedFilter, starred: starredFilter,
+  });
+  return res.json({
+    workspaces,
+    pagination: { total, limit: parsedLimit, offset: parsedOffset, hasMore: parsedOffset + workspaces.length < total },
+  });
+}
+
 // ---------------------------------------------------------------------------
 // GET /api/projects/:projectId/workspaces — list workspaces (root sessions)
 //
@@ -139,77 +193,7 @@ function handleCreateError(res, session, error, label) {
 //   Without `limit` query param → bare array of root session rows.
 //   With `limit` query param    → { workspaces: [...], pagination: { total, limit, offset, hasMore } }
 // ---------------------------------------------------------------------------
-projectWorkspacesRouter.get('/:projectId/workspaces', (req, res) => {
-  const startedAt = performance.now();
-  const project = projects.getById(req.params.projectId);
-  if (!project) {
-    return res.status(404).json({ error: ERR_PROJECT_NOT_FOUND });
-  }
-
-  const { archived, starred, limit, offset, view, status, scheduled } = req.query;
-
-  // The card view is the bounded hot-path used by the workspace list. Keep the
-  // original response below for compatibility with automation and integrations.
-  if (view === 'cards') {
-    const parsedLimit = Number.parseInt(limit, 10);
-    const parsedOffset = offset === undefined ? 0 : Number.parseInt(offset, 10);
-    if (!Number.isInteger(parsedLimit) || parsedLimit < 1 || parsedLimit > 50
-      || !Number.isInteger(parsedOffset) || parsedOffset < 0
-      || !['running', 'idle', undefined].includes(status)
-      || !['true', 'false', undefined].includes(scheduled)) {
-      return res.status(400).json({ error: 'Invalid workspace card pagination or filters' });
-    }
-    const pagePlusOne = sessions.getWorkspaceCards(req.params.projectId, {
-      archived: archived === 'true',
-      starred: starred === 'true' ? true : starred === 'false' ? false : null,
-      status: status || null,
-      scheduled: scheduled === 'true' ? true : scheduled === 'false' ? false : null,
-      limit: parsedLimit + 1,
-      offset: parsedOffset,
-    });
-    const hasMore = pagePlusOne.length > parsedLimit;
-    const cards = pagePlusOne.slice(0, parsedLimit);
-    return sendWorkspaceJson(res, {
-      workspaces: cards,
-      pagination: { limit: parsedLimit, offset: parsedOffset, hasMore },
-    }, startedAt);
-  }
-  let archivedFilter = null;
-  if (archived === 'true') archivedFilter = true;
-  else if (archived === 'false') archivedFilter = false;
-
-  let starredFilter = null;
-  if (starred === 'true') starredFilter = true;
-  else if (starred === 'false') starredFilter = false;
-
-  const parsedLimit = limit ? parseInt(limit, 10) : null;
-  const parsedOffset = offset ? parseInt(offset, 10) : 0;
-
-  const workspaces = sessions.getRootsByProjectId(req.params.projectId, {
-    archived: archivedFilter,
-    starred: starredFilter,
-    limit: parsedLimit,
-    offset: parsedOffset,
-  });
-
-  if (parsedLimit !== null) {
-    const total = sessions.getRootsCountByProjectId(req.params.projectId, {
-      archived: archivedFilter,
-      starred: starredFilter,
-    });
-    return res.json({
-      workspaces: workspaces.map(withPendingAgentInput),
-      pagination: {
-        total,
-        limit: parsedLimit,
-        offset: parsedOffset,
-        hasMore: parsedOffset + workspaces.length < total,
-      },
-    });
-  }
-
-  return res.json(workspaces.map(withPendingAgentInput));
-});
+projectWorkspacesRouter.get('/:projectId/workspaces', listProjectWorkspaces);
 
 // ---------------------------------------------------------------------------
 // POST /api/projects/:projectId/workspaces — create a new workspace
